requirements.txt
blinker==1.7.0
click==8.1.7
dominate==2.9.1
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.2
pandas==2.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.1
visitor==0.1.3
Werkzeug==3.0.1
gunicorn==21.2.0